Move webpayplus.py content to __init__.py
